+++ src/main/java/com/nbrichau/vanillaextension/init/TrapdoorInit.java
@@ -190,6 +190,7 @@ public class TrapdoorInit {
 	public static final Block purpur_block_trapdoor = null;
 	public static final Block purpur_pillar_trapdoor = null;
 	public static final Block end_stone_brick_trapdoor = null;
+	public static final Block grass_path_trapdoor = null;
 	public static final Block magma_block_trapdoor = null;
 	public static final Block nether_wart_block_trapdoor = null;
 	public static final Block red_nether_brick_trapdoor = null;
@@ -454,6 +455,7 @@ public class TrapdoorInit {
 		event.getRegistry().register(new TrapDoorBlock(Block.Properties.from(Blocks.PURPUR_BLOCK)).setRegistryName("purpur_block_trapdoor"));
 		event.getRegistry().register(new TrapDoorBlock(Block.Properties.from(Blocks.PURPUR_PILLAR)).setRegistryName("purpur_pillar_trapdoor"));
 		event.getRegistry().register(new TrapDoorBlock(Block.Properties.from(Blocks.END_STONE_BRICKS)).setRegistryName("end_stone_brick_trapdoor"));
+		event.getRegistry().register(new GrassPathTrapdoor(Block.Properties.from(Blocks.GRASS_PATH)).setRegistryName("grass_path_trapdoor"));
 		event.getRegistry().register(new TrapDoorBlock(Block.Properties.create(Material.ROCK, MaterialColor.NETHERRACK).setLightLevel((state)->{return 3;}).hardnessAndResistance(0.5F)).setRegistryName("magma_block_trapdoor"));
 		event.getRegistry().register(new TrapDoorBlock(Block.Properties.from(Blocks.NETHER_WART_BLOCK)).setRegistryName("nether_wart_block_trapdoor"));
 		event.getRegistry().register(new TrapDoorBlock(Block.Properties.from(Blocks.RED_NETHER_BRICKS)).setRegistryName("red_nether_brick_trapdoor"));
@@ -718,6 +720,7 @@ public class TrapdoorInit {
 		event.getRegistry().register(new BlockItem(purpur_block_trapdoor, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("purpur_block_trapdoor"));
 		event.getRegistry().register(new BlockItem(purpur_pillar_trapdoor, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("purpur_pillar_trapdoor"));
 		event.getRegistry().register(new BlockItem(end_stone_brick_trapdoor, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("end_stone_brick_trapdoor"));
+		event.getRegistry().register(new BlockItem(grass_path_trapdoor, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("grass_path_trapdoor"));
 		event.getRegistry().register(new BlockItem(magma_block_trapdoor, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("magma_block_trapdoor"));
 		event.getRegistry().register(new BlockItem(nether_wart_block_trapdoor, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("nether_wart_block_trapdoor"));
 		event.getRegistry().register(new BlockItem(red_nether_brick_trapdoor, new Item.Properties().group(VanillaExtension.VanillaExtensionItemGroup.instance)).setRegistryName("red_nether_brick_trapdoor"));