+++ src/main/java/com/nbrichau/vanillaextension/stairs/GrassBlockStairs.java
@@ -36,6 +36,13 @@ public class GrassBlockStairs extends StairsBlock {
 				player.getHeldItem(handIn).damageItem(1, player, item->item.sendBreakAnimation(handIn));
 				return ActionResultType.SUCCESS;
 			}
+			if (player.getHeldItem(handIn).getToolTypes().contains(ToolType.SHOVEL)) {
+				BlockState bs = StairsInit.grass_path_stairs.getDefaultState().with(FACING, state.get(FACING)).with(HALF, state.get(HALF)).with(SHAPE, state.get(SHAPE)).with(WATERLOGGED, state.get(WATERLOGGED));
+				worldIn.setBlockState(pos, bs);
+				worldIn.playSound(null,pos, SoundEvents.ITEM_SHOVEL_FLATTEN, SoundCategory.BLOCKS, 1.0F,1.0F);
+				player.getHeldItem(handIn).damageItem(1, player, item->item.sendBreakAnimation(handIn));
+				return ActionResultType.SUCCESS;
+			}
 		}
 		return super.onBlockActivated(state, worldIn, pos, player, handIn, hit);
 	}